Simulation/requirements.txt
# Core simulation dependencies
mujoco>=2.3.0
numpy>=1.21.0
glfw>=2.0.0

# Input handling
keyboard>=0.13.5

# Image processing for GUI
Pillow>=8.0.0

# Testing framework
pytest>=6.0.0
pytest-cov>=2.10.0

# Development tools
black>=21.0.0
flake8>=3.8.0
mypy>=0.910
